ASITiger: made joystick speed float
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13340 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/ASITiger/ASIPiezo.cpp
@@ -173,13 +173,13 @@ int CPiezo::Initialize()
 
    // joystick fast speed (JS X=) (per-card, not per-axis)
    pAct = new CPropertyAction (this, &CPiezo::OnJoystickFastSpeed);
-   CreateProperty(g_JoystickFastSpeedPropertyName, "100", MM::Integer, false, pAct);
+   CreateProperty(g_JoystickFastSpeedPropertyName, "100", MM::Float, false, pAct);
    UpdateProperty(g_JoystickFastSpeedPropertyName);
    SetPropertyLimits(g_JoystickFastSpeedPropertyName, 0, 100);
 
    // joystick slow speed (JS Y=) (per-card, not per-axis)
    pAct = new CPropertyAction (this, &CPiezo::OnJoystickSlowSpeed);
-   CreateProperty(g_JoystickSlowSpeedPropertyName, "10", MM::Integer, false, pAct);
+   CreateProperty(g_JoystickSlowSpeedPropertyName, "10", MM::Float, false, pAct);
    UpdateProperty(g_JoystickSlowSpeedPropertyName);
    SetPropertyLimits(g_JoystickSlowSpeedPropertyName, 0, 100);
