+++ DeviceAdapters/ASITiger/ASIXYStage.cpp
@@ -197,13 +197,13 @@ int CXYStage::Initialize()
 
    // joystick fast speed (JS X=)
    pAct = new CPropertyAction (this, &CXYStage::OnJoystickFastSpeed);
-   CreateProperty(g_JoystickFastSpeedPropertyName, "100", MM::Integer, false, pAct);
+   CreateProperty(g_JoystickFastSpeedPropertyName, "100", MM::Float, false, pAct);
    UpdateProperty(g_JoystickFastSpeedPropertyName);
    SetPropertyLimits(g_JoystickFastSpeedPropertyName, 0, 100);
 
    // joystick slow speed (JS Y=)
    pAct = new CPropertyAction (this, &CXYStage::OnJoystickSlowSpeed);
-   CreateProperty(g_JoystickSlowSpeedPropertyName, "10", MM::Integer, false, pAct);
+   CreateProperty(g_JoystickSlowSpeedPropertyName, "10", MM::Float, false, pAct);
    UpdateProperty(g_JoystickSlowSpeedPropertyName);
    SetPropertyLimits(g_JoystickSlowSpeedPropertyName, 0, 100);
